+++ srcs/array/macro/arr_size.c
@@ -6,13 +6,13 @@
 /*   By: ethebaul <ethebaul@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/21 20:27:56 by ethebaul          #+#    #+#             */
-/*   Updated: 2025/04/23 00:38:06 by ethebaul         ###   ########.fr       */
+/*   Updated: 2025/04/23 05:36:47 by ethebaul         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libc.h"
 
-uint	arr_size(void *array)
+ulong	arr_size(void *array)
 {
 	return (((t_array *)array)->size);
 }
